// src/app/services/model/employee.interface.ts
export interface Employee  {
     id: number;
     name: string;
     password: string;
     descr: string;
     gross_salary_month: number;
     health_insurance_share: number;
     surcharge: number;
     type: string;
     working_contract: string;
     contract_hours_month: number;
     limit_hours_month: number;
     ref_id: string;
     valid: number;
}